Rollup of include fixes for compatibility with modern compilers. Minimize warnings from redefinitions that differ by a space, fix headers that require other headers, add stdbool, stddef, and unistd.h, all of which are post-V7 but wanted by the compiler so badly that it pushes us to include them in userspace (not kernel -freestanding) so I sketched them in.

+++ usr/include/string.h
@@ -1,5 +1,7 @@
 // Not present in UNIXv7. long unsigned int deferred to size_t from ISO c89. RJL
 
+#include <stddef.h>
+
 void *   memchr (const void *, int, long unsigned int);
 int      memcmp (const void *, const void *, long unsigned int);
 void *   memcpy (void *__restrict, const void *__restrict, long unsigned int);
@@ -12,7 +14,7 @@ int      strcoll (const char *, const char *);
 char    *strcpy (char *__restrict, const char *__restrict);
 long unsigned int   strcspn (const char *, const char *);
 char    *strerror (int);
-long unsigned int strlen (const char *);
+size_t strlen (const char *);
 char    *strncat (char *__restrict, const char *__restrict, long unsigned int);
 int      strncmp (const char *, const char *, long unsigned int);
 char    *strncpy (char *__restrict, const char *__restrict, long unsigned int);
